# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "MediaItem": {
        "properties": {
          "contentType": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "height": {
            "format": "int32",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "size": {
            "format": "int64",
            "type": "integer"
          },
          "width": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "MemberResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ReportClaim": {
        "properties": {
          "classification": {
            "minLength": 1,
            "pattern": "DOMESTIC|IMPORTED_SPECIFIED|IMPORTED_UNSPECIFIED|MIXED|UNKNOWN",
            "type": "string"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/ReportOriginComponent"
            },
            "maxItems": 20,
            "minItems": 0,
            "type": "array"
          },
          "ingredientId": {
            "format": "uuid",
            "type": "string"
          },
          "originalExpression": {
            "maxLength": 1000,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "classification",
          "components",
          "ingredientId",
          "originalExpression"
        ],
        "type": "object"
      },
      "ReportListItem": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "restaurantId": {
            "format": "uuid",
            "type": "string"
          },
          "restaurantName": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "ReportOriginComponent": {
        "properties": {
          "countryCode": {
            "pattern": "[A-Z]{2}",
            "type": "string"
          },
          "kind": {
            "minLength": 1,
            "pattern": "DOMESTIC|IMPORTED_SPECIFIED|IMPORTED_UNSPECIFIED",
            "type": "string"
          },
          "ratio": {
            "exclusiveMinimum": 0,
            "maximum": 1,
            "type": "number"
          }
        },
        "required": [
          "kind"
        ],
        "type": "object"
      },
      "ReportPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReportListItem"
            },
            "type": "array"
          },
          "nextCursor": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
      "ReportReviewHistory": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "decision": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ReportSubmission": {
        "properties": {
          "claims": {
            "items": {
              "$ref": "#/components/schemas/ReportClaim"
            },
            "maxItems": 10,
            "minItems": 1,
            "type": "array"
          },
          "mediaIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 5,
            "minItems": 1,
            "type": "array"
          },
          "observedOn": {
            "format": "date",
            "type": "string"
          },
          "publicationConsent": {
            "type": "boolean"
          },
          "restaurantId": {
            "format": "uuid",
            "type": "string"
          },
          "scopeId": {
            "format": "uuid",
            "type": "string"
          },
          "scopeName": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "usage": {
            "minLength": 1,
            "pattern": "SIDE_DISH|STEW|MAIN_DISH|OTHER|UNKNOWN",
            "type": "string"
          }
        },
        "required": [
          "claims",
          "mediaIds",
          "observedOn",
          "restaurantId",
          "scopeName",
          "usage"
        ],
        "type": "object"
      },
      "ReportUpdateRequest": {
        "properties": {
          "expectedVersion": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "submission": {
            "$ref": "#/components/schemas/ReportSubmission"
          }
        },
        "required": [
          "expectedVersion",
          "submission"
        ],
        "type": "object"
      },
      "ReportView": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ReportReviewHistory"
            },
            "type": "array"
          },
          "state": {
            "type": "string"
          },
          "submission": {
            "$ref": "#/components/schemas/ReportSubmission"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "ReportWithdrawRequest": {
        "properties": {
          "expectedVersion": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "reason": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "expectedVersion",
          "reason"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "serviceBearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "국산김치맵 API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/reports/{id}": {
      "get": {
        "operationId": "detail_2",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReportView"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "tags": [
          "admin-report-controller"
        ]
      }
    },
...
    "/api/v1/media": {
      "post": {
        "operationId": "upload",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaItem"
                }
              }
            },
            "description": "Created"
          }
        },
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "tags": [
          "media-controller"
        ]
      }
    },
    "/api/v1/media/{id}": {
      "get": {
        "operationId": "download",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "original",
            "required": false,
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "format": "byte",
                  "type": "string"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "tags": [
          "media-controller"
        ]
      }
    },
    "/api/v1/media/{id}/public": {
      "get": {
        "operationId": "publicImage",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "format": "byte",
                  "type": "string"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "media-controller"
        ]
      }
    },
    "/api/v1/members/me": {
      "delete": {
        "operationId": "withdraw_1",
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "tags": [
          "auth-controller"
        ]
      },
      "get": {
        "operationId": "me",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "tags": [
          "auth-controller"
        ]
      }
    },
    "/api/v1/reports": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReportPage"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "tags": [
          "report-controller"
        ]
      },
      "post": {
        "operationId": "create",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReportView"
                }
              }
            },
            "description": "Created"
          }
        },
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "tags": [
          "report-controller"
        ]
      }
    },
    "/api/v1/reports/{id}": {
      "get": {
        "operationId": "detail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReportView"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "tags": [
          "report-controller"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReportView"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "tags": [
          "report-controller"
        ]
      }
    },
    "/api/v1/reports/{id}/withdraw": {
      "post": {
        "operationId": "withdraw",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportWithdrawRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReportView"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "tags": [
          "report-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}